+++ modules/index-all.js
@@ -1,0 +1,2 @@
+export { default } from './index-default';
+export * from './index';
