--- underscore-umd.js
@@ -1,1 +1,0 @@
-export { default } from './underscore-module';
